Use scraper (html5ever) to parse (and unescape) page titles
Use a real HTML parsing library, which should give better resilience than a
simple regex.

This also side-steps issues seen with the `htmlescape` crate which is also now
unmaintained.

Change the page download code so that it now builds up the page content with
downloaded chunks until a title is found. This way, the smallest amount of the
page (by quantum of chunk size) is downloaded in order to obtain a title, while
pages with large amounts of gumpf (e.g. CSS) preceding the title have enough
content downloaded.